moved data access file to models folder
